move into target goal after move action is complete
