+++ src/components/GoalsComponents/MoveGoalButton.tsx
@@ -6,17 +6,21 @@ import { moveGoalHierarchy } from "@src/helpers/GoalController";
 import ZButton from "@src/common/ZButton";
 import { ILocationState } from "@src/Interfaces";
 import { GoalItem } from "@src/models/GoalItem";
+import useNavigateToSubgoal from "@src/store/useNavigateToSubgoal";
 
 interface GoalMoveButtonProps {
   type: "select" | "move";
   goal?: GoalItem;
-  targetGoalId?: string;
+  targetGoal?: GoalItem;
 }
 
-const GoalMoveButton: React.FC<GoalMoveButtonProps> = ({ goal, targetGoalId, type }) => {
+const GoalMoveButton: React.FC<GoalMoveButtonProps> = ({ goal, targetGoal, type }) => {
   const navigate = useNavigate();
   const { state }: { state: ILocationState } = useLocation();
   const [selectedGoal, setSelectedGoal] = useRecoilState(moveGoalState);
+  const navigateToSubgoal = useNavigateToSubgoal();
+
+  const targetGoalId = targetGoal?.id;
 
   const handleClick = () => {
     if (type === "select") {
@@ -27,6 +31,7 @@ const GoalMoveButton: React.FC<GoalMoveButtonProps> = ({ goal, targetGoalId, typ
     } else if (type === "move") {
       if (selectedGoal && targetGoalId) {
         moveGoalHierarchy(selectedGoal, targetGoalId);
+        navigateToSubgoal(targetGoal);
         setSelectedGoal(null);
       }
     }
